+++ NEWS.md
@@ -6,6 +6,7 @@
 ## Minor changes
 - `geom_half_point` now supports passing of positions other than `PositionJitter` and `PositionIdentity` to `transformation` -- given they have a `compute_panel` function.
 - added `range_scale` parameter to `geom_half_point`. When no `width` argument is passed in `transformation_params`, `range_scale` defaults to `0.5`, which leads to the points using half of their allotted space. `1` would have them end up using all of their allotted space. 
+- added default `width` and `height` values when `width` or `height` are included names in the `transformation_params` list but set to `NULL`. This leads to sensible defaults for `Position`s that are not `PositionJitter`. 
 
 
 ## Bug Fixes